Fixed some WriteJson calls in converters to ease unit testing

+++ CryptoExchange.Net/Converters/TimestampConverter.cs
@@ -21,7 +21,7 @@ namespace CryptoExchange.Net.Converters
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            writer.WriteValue(Math.Round((((DateTime)value) - new DateTime(1970, 1, 1)).TotalMilliseconds));
+            writer.WriteValue((long)Math.Round((((DateTime)value) - new DateTime(1970, 1, 1)).TotalMilliseconds));
         }
     }
 }
